Restructure src directory and fix Docker builds
