+++ dvc.lock
@@ -1,39 +1,44 @@
 schema: '2.0'
 stages:
   preprocess:
-    cmd: python3 src/run.py --preprocess
+    cmd: python3 src/preprocess_data.py
     deps:
     - path: data/
       md5: 3d9e62a9c241fcfcf634384a95dd4cf4.dir
       size: 10408878
       nfiles: 4
+    - path: src/common/
+      md5: 5539f7e5a5514497fe76383475197425.dir
+      size: 3115
+      nfiles: 4
     - path: src/preprocess_data.py
-      md5: 4418185e1646a2c3c53540b858582de9
-      size: 4177
+      md5: b1e680538f0a6e8e3026c983335b2807
+      size: 3414
     outs:
     - path: outputs/models/vectorizer.pkl
-      md5: 3d153c9e0ca129961bc4477e3c73bea9
+      md5: 0a8b5f842d4c8de8c55507479f1fddb1
       size: 6039557
     - path: outputs/processed_data/
       md5: 7730ae7add1883a1d89e24abfc316802.dir
       size: 112774922
       nfiles: 6
   train:
-    cmd: python3 src/run.py --train
+    cmd: python3 src/train_model.py
     deps:
     - path: outputs/models/vectorizer.pkl
-      md5: 3d153c9e0ca129961bc4477e3c73bea9
+      md5: 0a8b5f842d4c8de8c55507479f1fddb1
       size: 6039557
     - path: outputs/processed_data/
       md5: 7730ae7add1883a1d89e24abfc316802.dir
       size: 112774922
       nfiles: 6
-    - path: src/preprocess_data.py
-      md5: 4418185e1646a2c3c53540b858582de9
-      size: 4177
+    - path: src/common/
+      md5: 5539f7e5a5514497fe76383475197425.dir
+      size: 3115
+      nfiles: 4
     - path: src/train_model.py
-      md5: 823fa8ae9ec3737ee04a0939eea0f37b
-      size: 1003
+      md5: cc0c0b52fca0cf4bc6330f5da833b6f8
+      size: 1265
     outs:
     - path: outputs/models/tfidf_model.pkl
       md5: 40c023aa76368aa506f19c1ed2a9ef45
@@ -42,8 +47,15 @@ stages:
     cmd: docker build -t nidhogg-frontend -f docker/frontend.Dockerfile .
     deps:
     - path: docker/frontend.Dockerfile
-      md5: 898eb5733e6be7a83d1baabd1c3967ff
-      size: 558
+      md5: 95444a2db35eebda32ff78852a3e7d37
+      size: 586
+    - path: src/frontend.py
+      md5: 971b79df95933a8eb9bfbf73238dfa44
+      size: 2443
+    - path: src/frontend/
+      md5: 49eccedec2f0dc1d9a354761129cae05.dir
+      size: 797
+      nfiles: 1
   build_backend:
     cmd: docker build -t nidhogg-backend -f inference/Dockerfile .
     deps:
@@ -66,17 +78,18 @@ stages:
     cmd: docker build -t nidhogg-inference -f docker/inference.Dockerfile .
     deps:
     - path: docker/inference.Dockerfile
-      md5: fbcb7d09dadfce1c520e4b1359bc2260
-      size: 690
+      md5: 61df04af4be564fe6b76dd2bd847c1fc
+      size: 685
     - path: outputs/models/tfidf_model.pkl
       md5: 40c023aa76368aa506f19c1ed2a9ef45
       size: 21526649
     - path: outputs/models/vectorizer.pkl
-      md5: 3d153c9e0ca129961bc4477e3c73bea9
+      md5: 0a8b5f842d4c8de8c55507479f1fddb1
       size: 6039557
+    - path: src/common/
+      md5: 5539f7e5a5514497fe76383475197425.dir
+      size: 3115
+      nfiles: 4
     - path: src/inference_api.py
-      md5: 4954f2968a536938f78afbc619ed086b
-      size: 2447
-    - path: src/preprocess_data.py
-      md5: 4418185e1646a2c3c53540b858582de9
-      size: 4177
+      md5: e6bf0c90798901539b8dcbf17eac2778
+      size: 2732
