+++ dvc.yaml
@@ -1,16 +1,17 @@
 stages:
   preprocess:
-    cmd: python3 src/run.py --preprocess
+    cmd: python3 src/preprocess_data.py
     deps:
-    - data/
     - src/preprocess_data.py
+    - src/common/
+    - data/
     outs:
     - outputs/processed_data/
     - outputs/models/vectorizer.pkl
   train:
-    cmd: python3 src/run.py --train
+    cmd: python3 src/train_model.py
     deps:
-    - src/preprocess_data.py
+    - src/common/
     - src/train_model.py
     - outputs/processed_data/
     - outputs/models/vectorizer.pkl
@@ -20,11 +21,13 @@ stages:
     cmd: docker build -t nidhogg-inference -f docker/inference.Dockerfile .
     deps:
     - docker/inference.Dockerfile
-    - src/preprocess_data.py
     - src/inference_api.py
+    - src/common/
     - outputs/models/vectorizer.pkl
     - outputs/models/tfidf_model.pkl
   build_frontend:
     cmd: docker build -t nidhogg-frontend -f docker/frontend.Dockerfile .
     deps:
     - docker/frontend.Dockerfile
+    - src/frontend.py
+    - src/frontend/
